requirements.txt
flask==2.3.3
tensorflow==2.12.0
pillow==9.4.0
numpy==1.23.5